Re-organize the integration tests for RTMClient
